Move slurm fns to slurm.py; remove from filesystem.py
